# Conflicts:
#	app/src/main/java/com/xjek/provider/views/dashboard/DashBoardActivity.kt
#	base/src/main/java/com/xjek/base/data/Constants.kt
#	base/src/main/java/com/xjek/base/data/PreferencesKey.kt
#	taxiservice/src/main/java/com/xjek/taxiservice/views/main/TaxiDashboardActivity.kt

+++ app/src/main/java/com/xjek/provider/models/CheckRequestModel.kt
@@ -212,7 +212,7 @@ data class Invoice(
 
 data class Payment(
         var card: Int = 0,
-        var cash: Int = 0,
+        var cash: Double = 0.0,
         var commision: Double = 0.0,
         var commision_percent: Int = 0,
         var company_id: Int = 0,
